Tried OpenEars API for offline voice recognition.
Implemented voice recognition using the OpenEars API for recognition
when the user is offline.
Voice recognition accuracy was very low, so I decided it was better to
provide no recognition than to provide terrible results.
Therefore, committing this code to a separate branch, which should not
be merged into master.